Fixed another bug in admin_domains
git-svn-id: file:///var/svn/trunk@170 45fdb5c4-e40b-0410-b369-9aab4fe9a275

+++ syscp/admin_domains.php
@@ -102,10 +102,17 @@
 						$idString[] = '`domainid` = "'.$subRow['id'].'"';	
 					}
 					$idString = implode(' OR ', $idString);
-					$query = 
-						'DELETE FROM `'.TABLE_MAIL_USERS.'` ' .
-						'WHERE '.$idString;
-					$db->query($query);
+					if ( $idString != '' )
+					{
+						$query = 
+							'DELETE FROM `'.TABLE_MAIL_USERS.'` ' .
+							'WHERE '.$idString;
+						$db->query($query);
+						$query = 
+							'DELETE FROM `'.TABLE_MAIL_VIRTUAL.'` ' .
+							'WHERE '.$idString;
+						$db->query($query);
+					}
 					$db->query("DELETE FROM `".TABLE_PANEL_DOMAINS."` WHERE `id`='$id' OR `parentdomainid`='".$result['id']."'");
 					$deleted_domains = $db->affected_rows();
 					$db->query("UPDATE `".TABLE_PANEL_CUSTOMERS."` SET `subdomains_used` = `subdomains_used` - 0".($deleted_domains - 1)." WHERE `customerid` = '{$result['customerid']}'");
